+++ components/tag/index.ts
@@ -1,5 +1,5 @@
 import Tag from './Tag.vue'
-export type { Props, Tag } from './Tag.vue'
+export type { Props, Item } from './Tag.vue'
 import { withInstall } from '../utils/type'
 
 export default withInstall(Tag)